+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Yii Strings Change Log
 
-
 ## 2.0.1 under development
 
-- New #75: Add method `Inflector::toSnakeCase()` that convert word to "snake_case" (soodssr)
+- Enh #75: Add method `Inflector::toSnakeCase()` that convert word to "snake_case" (soodssr)
+- Enh: Add method `NumericHelper:isPositiveInteger()` that checks whether the given string is a positive integer number 
+(arogachev)
 
 ## 2.0.0 February 10, 2021
 
@@ -17,7 +18,7 @@
 ## 1.2.0 January 22, 2021
 
 - Enh #62: Add method `StringHelper::split()` that split a string to array with non-empty lines (vjik)
-- Enh #63: Add method `NumericHelper::isInteger()` that checks whether the given string is an integer number (vjik)
+- Enh #63: Add method `NumericHelper::isInteger()` that checks whether the given value is an integer number (vjik)
 - Enh #64: Add support of a boolean values to `NumericHelper::normalize()` (vjik)
 
 ## 1.1.0 November 13, 2020